requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
class-registry==2.1.2
cloudinary==1.41.0
dj-database-url==2.2.0
Django==4.1.13
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-mongoengine==0.5.6
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.6.1
filters==1.3.2
gunicorn==22.0.0
idna==3.8
mongoengine==0.28.2
packaging==24.0
pillow==10.4.0
psycopg2-binary==2.9.9
pymongo==3.12.3
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
regex==2024.5.10
requests==2.32.3
six==1.16.0
sqlparse==0.2.4
typing_extensions==4.12.0
urllib3==2.2.2
whitenoise==6.8.2
